backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-docx==0.8.11
pydantic==1.10.12
email-validator==1.3.1
Flask==2.2.5
flask-cors==3.0.10
google-generativeai
python-docx
reportlab
